RDM-3230 Updates necessary for Heroku.
